feat: 사용자 별점 기능을 위한 사용자 평점 렌더링 로직을 userRatingView로 이동 및 스타일 수정
